backend/requirements.txt
# Django and REST Framework
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0

# JWT Authentication
djangorestframework-simplejwt==5.4.0

# Database
mysqlclient==2.2.6
python-dotenv==1.0.1

# Production
gunicorn==23.0.0
whitenoise==6.8.2

# Async & WebSockets (New)
daphne>=4.0.0
channels>=4.0.0
channels_redis>=4.0.0

# Development and Testing
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

# Utilities
Pillow==11.0.0
django-filter==24.3
